fix: refactor actor/casting JobListPage to component
