+++ apps/web/src/modules/job/list/components/ActorCastingPage/placeholder.ts
